refactor(backend): BBB api types in separate file
