+++ backend/src/api/BBB.ts
@@ -4,70 +4,17 @@ import { CONFIG } from '#config/config'
 import logger from '#src/logger'
 
 import { axiosInstance, createChecksum } from './BBB/axios'
+import {
+  CreateMeetingResponse,
+  MeetingInfo,
+  GetMeetingsResponse,
+  CreateMeetingOptions,
+  JoinMeetinLinkOptions,
+} from './BBB/types'
+
+export { MeetingInfo, AttendeeInfo } from './BBB/types'
 
 const parser = new XMLParser()
-
-interface CreateMeetingResponse {
-  returncode: string
-  meetingID: string
-  internalMeetingID: string
-  parentMeetingID: string
-  attendeePW: string
-  moderatorPW: string
-  createTime: number
-  voiceBridge: number
-  dialNumber: string
-  createDate: Date
-  hasUserJoined: boolean
-  duration: number
-  hasBeenForciblyEnded: boolean
-  messageKey: string
-  message: string
-}
-
-export type AttendeeInfo = {
-  userID: string
-  fullName: string
-  role: string
-  isPresenter: boolean
-  isListeningOnly: boolean
-  hasJoinedVoice: boolean
-  hasVideo: boolean
-  clientType: string
-}
-
-export type MeetingInfo = {
-  meetingName: string
-  meetingID: string
-  internalMeetingID: string
-  createTime: number
-  createDate: Date
-  voiceBridge: number
-  dialNumber: string
-  attendeePW: string
-  moderatorPW: string
-  running: boolean
-  duration: number
-  hasUserJoined: boolean
-  recording: boolean
-  hasBeenForciblyEnded: boolean
-  startTime: number
-  endTime: number
-  participantCount: number
-  listenerCount: number
-  voiceParticipantCount: number
-  videoCount: number
-  maxUsers: number
-  moderatorCount: number
-  attendees: string | { attendee: AttendeeInfo[] | AttendeeInfo }
-  metadata: string
-  isBreakout: boolean
-}
-
-type GetMeetingsResponse = {
-  returncode: string
-  meetings: { meeting: MeetingInfo[] } | string
-}
 
 export const getMeetings = async (): Promise<MeetingInfo[]> => {
   try {
@@ -91,12 +38,6 @@ export const getMeetings = async (): Promise<MeetingInfo[]> => {
     logger.error('getMeetings with error', err)
   }
   return []
-}
-
-interface CreateMeetingOptions {
-  name: string
-  meetingID: string
-  // welcome?: string
 }
 
 export const createMeeting = async (
@@ -125,15 +66,6 @@ export const createMeeting = async (
     logger.error('createMeeting with error', err)
     return null
   }
-}
-
-interface JoinMeetinLinkOptions {
-  fullName: string
-  meetingID: string
-  role?: 'MODERATOR' | 'VIEWER'
-  password?: string
-  createTime?: string
-  userID?: string
 }
 
 export const joinMeetingLink = (options: JoinMeetinLinkOptions): string => {
